% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{build_sum_tab}
\alias{build_sum_tab}
\title{\code{\link{summary.em.fit}} helper: build summary table}
\usage{
build_sum_tab(x, indent = 4L)
}
\arguments{
\item{x}{a \code{data.frame} or \code{\link[tibble]{tibble}} with numeric columns, one for each label class}

\item{indent}{integer. No. white-space indentations added to the left (default: 4)}
}
\value{
a scalar character vector
}
\description{
\code{\link{summary.em.fit}} helper: build summary table
}
