% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{Estimate Dawid-Skene Expectation Maximization model for categorical annotation data}
\usage{
em(data, item.col, annotator.col, label.col, max.iters = 100,
  .ability.prior = 0.7, .prevalence.prior = NULL, .beta.prior = 0.01,
  .alpha.prior = 0.01, .min.relative.diff = 1e-08, verbose = TRUE)
}
\arguments{
\item{data}{a data frame with rows corresponding to annotator \eqn{j}'s annotation \eqn{y_{ij}} of item \eqn{i}.}

\item{item.col}{(unquoted name of) column containing item identifiers}

\item{annotator.col}{(unquoted name of) column containing annotator identifiers}

\item{label.col}{(unquoted name of) column containing label identifiers}

\item{max.iters}{positive integer, specifying the maximum number of EM iterations to perform.
(CAUTION: low values may cause stopping before convergence)
Defaults to 100.}

\item{.ability.prior}{number in interval (0, 1), specifying the annotator ability prior used to initialize ability parameters for first iteration
Defaults to 0.7}

\item{.prevalence.prior}{vector of numbers in interval (0, 1),
specifying prior probabilities on item label classes \eqn{\hat{\boldsymbol{pi}}}.
Most have length equal to number of label classes (i.e., number of unique item labels in \code{data}).
Defaults to \code{NULL}, in which case prior class label probabilities are set all equal to \eqn{1/K} (where \eqn{K} is the number of item label classes)}

\item{.beta.prior}{number in interval (0, 1), specifying the beta prior used to perform smoothing on \eqn{\hat{\boldsymbol{pi}}} (label class prevalence estimates)
Defaults to 0.01}

\item{.alpha.prior}{number in interval (0, 1), specifying the alpha prior used to perform smoothing on items' label class probability estimates
Defaults to 0.01}

\item{.min.relative.diff}{number in interval (0, 1), specifying the minimum relative difference between log-likelihoods of two subsequent iteration steps.
Governs when convergence is reached (the lower, the more iterations).
Defaults to 1e-8}

\item{verbose}{logical. print-out data description and iteration history?}
}
\description{
Function fits an Dawid-Skene annotation model to categorical annotation data
    using the Expectation Maximization algorithm.
}
\note{
Code adapted from Bob Carpenter's implementation
   of Dawid and Skene (1979) "Maximum Likelihood Estimation of Observer Error-Rates Using the EM Algorithm"
   (see \url{https://github.com/bob-carpenter/anno/blob/master/R/em-dawid-skene.R})
}
