% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.bin.R
\name{emplot_annotator_ability_distributions}
\alias{emplot_annotator_ability_distributions}
\title{EM-fit plot: Annotator ability estimate distributions}
\usage{
emplot_annotator_ability_distributions(em.fit)
}
\arguments{
\item{em.fit}{an \code{em.fit} object es returned by \code{em}}
}
\value{
a \code{em.fit.plot} object inheriting from \code{\link[ggplot2]{ggplot}}
}
\description{
EM-fit plot: Annotator ability estimate distributions
}
