% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.bin.R
\name{emplot_convergence}
\alias{emplot_convergence}
\title{EM-fit plot: estimation convergence}
\usage{
emplot_convergence(em.fit)
}
\arguments{
\item{em.fit}{an \code{em.fit} object es returned by \code{em}}
}
\value{
a \code{em.fit.plot} object inheriting from \code{\link[ggplot2]{ggplot}}
}
\description{
EM-fit plot: estimation convergence
}
