% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{emplot_item_class_distributions}
\alias{emplot_item_class_distributions}
\title{EM-fit plot: Item class probability estimate distributions}
\usage{
emplot_item_class_distributions(em.fit)
}
\arguments{
\item{em.fit}{an \code{em.fit} object es returned by \code{em}}
}
\value{
a \code{em.fit.plot} object inheriting from \code{\link[ggplot2]{ggplot}}
}
\description{
EM-fit plot: Item class probability estimate distributions
}
