% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.bin.R
\name{get_annotations_info}
\alias{get_annotations_info}
\title{Get annotations info}
\usage{
get_annotations_info(x)
}
\arguments{
\item{x}{a \code{data.frame} with columns '_item', '_annotator' and '_label'
holding corresponding integer indexes}
}
\value{
a list
}
\description{
Get annotations info
}
