% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_data_info}
\alias{get_data_info}
\title{\code{\link{summary.em.fit}} helper: Get data info}
\usage{
get_data_info(object)
}
\arguments{
\item{object}{an object of class "em.fit"}
}
\value{
a scalar character vector
}
\description{
\code{\link{summary.em.fit}} helper: Get data info
}
