% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_iter_info}
\alias{get_iter_info}
\title{\code{\link{summary.em.fit}} helper: Get iterations info}
\usage{
get_iter_info(object)
}
\arguments{
\item{object}{an object of class "em.fit"}
}
\value{
a scalar character vector
}
\description{
\code{\link{summary.em.fit}} helper: Get iterations info
}
