% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_lab_sum}
\alias{get_lab_sum}
\title{\code{\link{summary.em.fit}} helper: Get label summary statistics}
\usage{
get_lab_sum(lab)
}
\arguments{
\item{lab}{an numeric vector of class label estimates}
}
\value{
a list
}
\description{
\code{\link{summary.em.fit}} helper: Get label summary statistics
}
