% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_labeling}
\alias{get_labeling}
\title{Get labeling}
\usage{
get_labeling(object)
}
\arguments{
\item{object}{the model object}
}
\value{
a \code{\link[tibble]{tibble}} mapping items to class category estimates
}
\description{
Given a model object, function induces a labeling from class category probability estimates.
}
