% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.bin.R
\name{init_em_bin_params}
\alias{init_em_bin_params}
\title{Initialize EM-parameters for binar model}
\usage{
init_em_bin_params(x, .ability.prior)
}
\arguments{
\item{x}{a list as returned by \code{\link{create_em_bin_indixes}}}

\item{.ability.prior}{non-negative number, specifying the annotator ability prior used to initialize ability parameters for first iteration
Defaults to 0.7}
}
\value{
a list with elements
    \itemize{
      \item 'theta_hat': annotator parameter matrices (a 3d-tensor with extension \eqn{K* K* J})
      \item 'pi_hat': prevalence estimates (a \eqn{K}-length numeric vector)
      \item 'E_z': \eqn{I* K} matrix of item label class probability estimates
    }
}
\description{
Initialize EM-parameters for binar model
}
