% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{pad_ws}
\alias{pad_ws}
\title{Pad white space}
\usage{
pad_ws(x, n, left = TRUE)
}
\arguments{
\item{x}{character vector}

\item{n}{total No. characters}

\item{left}{padding to the left (default: \code{TRUE}). otherwise padding to the right}
}
\value{
a character vector
}
\description{
Pad white space
}
