% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.em.fit}
\alias{print.summary.em.fit}
\title{S3 method for class "summary.em.fit"}
\usage{
\method{print}{summary.em.fit}(x, ...)
}
\arguments{
\item{x}{an object of class "summary.em.fit"}

\item{...}{additional arguments are ignored}
}
\value{
invisible
}
\description{
S3 method for class "summary.em.fit"
}
