% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{rebuild_call}
\alias{rebuild_call}
\title{\code{\link{summary.em.fit}} helper: Rebuild call}
\usage{
rebuild_call(object)
}
\arguments{
\item{object}{an object of class "em.fit"}
}
\value{
a list
}
\description{
\code{\link{summary.em.fit}} helper: Rebuild call
}
