% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sprintf_lab_sum}
\alias{sprintf_lab_sum}
\title{\code{\link{summary.em.fit}} helper: Printer for label summary statistics}
\usage{
sprintf_lab_sum(nm, vals)
}
\arguments{
\item{nm}{label class name}

\item{vals}{label class summary statistics}
}
\value{
a scalar character vector
}
\description{
\code{\link{summary.em.fit}} helper: Printer for label summary statistics
}
