% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.em.fit}
\alias{summary.em.fit}
\title{S3 method for class "summary.em.fit"}
\usage{
\method{summary}{em.fit}(object)
}
\arguments{
\item{object}{an object of class "em.fit", usually a result of \code{\link{em.bin}}}
}
\value{
an object of class "summary.em.fit"
}
\description{
summary method for class "em.fit"
}
