% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.bin.R
\name{em}
\alias{em}
\title{Estimate Dawid-Skene Expectation Maximization model for categorical annotation data}
\usage{
em(data, item.col, annotator.col, label.col, max.iters = 100,
  .ability.prior = 0.7, .beta.prior = 0.01, .alpha.prior = 0.01,
  .min.relative.diff = 1e-08, verbose = TRUE)
}
\arguments{
\item{data}{a data frame with rows corresponding to annotator \eqn{j}'s annotation \eqn{y_{ij}} of item \eqn{i}.}

\item{item.col}{(unquoted name of) column containing item identifiers}

\item{annotator.col}{(unquoted name of) column containing annotator identifiers}

\item{label.col}{(unquoted name of) column containing label identifiers}

\item{max.iters}{positive integer, specifying the maximum number of EM iterations to perform.
(CAUTION: low values may cause stopping before convergence)
Defaults to 100.}

\item{.ability.prior}{non-negative number, specifying the annotator ability prior used to initialize ability parameters for first iteration
Defaults to 0.7}

\item{.beta.prior}{non-negative number, specifying the beta prior used to perform smoothing on \eqn{\hat{pi}} (label class prevalence estimates)
Defaults to 0.01}

\item{.alpha.prior}{non-negative number, specifying the alpha prior used to perform smoothing on items' label class probability estimates
Defaults to 0.01}

\item{.min.relative.diff}{non-negative number, specifying the minimum relative difference between log-likelihoods of two subsequent iteration steps.
Governs when convergence is reached (the lower, the more iterations).
Defaults to 1e-8}

\item{verbose}{logical. print-out data description and iteration history?}
}
\description{

}
\note{
Code adapted from Bob Carpenter's implementation of
   of Dawid and Skene (1979) "Maximum Likelihood Estimation
   (see \url{https://github.com/bob-carpenter/anno/blob/master/R/em-dawid-skene.R})
}
